crypto-forecasting/requirements.txt
pandas==2.2.3
numpy==2.0.2
ta==0.11.0
matplotlib==3.9.4
neuralforecast==2.0.1
pytest==8.3.5 
python-dotenv==1.0.1